requirements.txt
discord.py
aiohttp
edge-tts
duckduckgo-search
python-dotenv